Gamma protocol adapter
