+++ projects/opyn/abis/gamma/isWhitelistedCollateral.json
@@ -1,0 +1,20 @@
+{
+  "inputs": [
+    {
+      "internalType": "address",
+      "name": "_collateral",
+      "type": "address"
+    }
+  ],
+  "name": "isWhitelistedCollateral",
+  "outputs": [
+    {
+      "internalType": "bool",
+      "name": "",
+      "type": "bool"
+    }
+  ],
+  "stateMutability": "view",
+  "type": "function",
+  "constant": true
+}
